+++ apollo-router/src/plugins/telemetry/formatters/mod.rs
@@ -24,8 +24,8 @@ use tracing_subscriber::registry::SpanRef;
 
 use super::config_new::logging::RateLimit;
 use super::dynamic_attribute::LogAttributes;
-use super::reload::SampledSpan;
 use crate::plugins::telemetry::otel::OtelData;
+use crate::plugins::telemetry::reload::otel::SampledSpan;
 
 pub(crate) const APOLLO_PRIVATE_PREFIX: &str = "apollo_private.";
 // FIXME: this is a temporary solution to avoid exposing hardcoded attributes in connector spans instead of using the custom telemetry features.
