+++ apollo-router/src/plugins/telemetry/metrics/apollo/mod.rs
@@ -10,7 +10,6 @@ use crate::plugins::telemetry::apollo::OtlpMetricsBatchProcessorConfiguration;
 use crate::plugins::telemetry::apollo::router_id;
 use crate::plugins::telemetry::apollo_exporter::ApolloExporter;
 use crate::plugins::telemetry::apollo_exporter::get_uname;
-use crate::plugins::telemetry::builder::MetricsBuilder;
 use crate::plugins::telemetry::config::{ApolloMetricsReferenceMode, Conf};
 use crate::plugins::telemetry::metrics::CustomAggregationSelector;
 use crate::plugins::telemetry::metrics::MetricsConfigurator;
@@ -18,6 +17,7 @@ use crate::plugins::telemetry::otlp::CustomTemporalitySelector;
 use crate::plugins::telemetry::otlp::Protocol;
 use crate::plugins::telemetry::otlp::TelemetryDataKind;
 use crate::plugins::telemetry::otlp::process_endpoint;
+use crate::plugins::telemetry::reload::builder::MetricsBuilder;
 use opentelemetry::KeyValue;
 use opentelemetry_otlp::MetricsExporterBuilder;
 use opentelemetry_otlp::WithExportConfig;
