+++ apollo-router/src/plugins/telemetry/metrics/otlp.rs
@@ -1,9 +1,9 @@
 use crate::metrics::aggregation::MeterProviderType;
-use crate::plugins::telemetry::builder::MetricsBuilder;
 use crate::plugins::telemetry::config::Conf;
 use crate::plugins::telemetry::metrics::CustomAggregationSelector;
 use crate::plugins::telemetry::metrics::MetricsConfigurator;
 use crate::plugins::telemetry::otlp::TelemetryDataKind;
+use crate::plugins::telemetry::reload::builder::MetricsBuilder;
 use opentelemetry_otlp::MetricsExporterBuilder;
 use opentelemetry_sdk::metrics::PeriodicReader;
 use opentelemetry_sdk::runtime;
