+++ apollo-router/src/plugins/telemetry/tracing/apollo.rs
@@ -6,9 +6,9 @@ use tower::BoxError;
 use crate::plugins::telemetry::apollo::Config;
 use crate::plugins::telemetry::apollo::router_id;
 use crate::plugins::telemetry::apollo_exporter::proto::reports::Trace;
-use crate::plugins::telemetry::builder::TracingBuilder;
 use crate::plugins::telemetry::config::Conf;
 use crate::plugins::telemetry::otel::named_runtime_channel::NamedTokioRuntime;
+use crate::plugins::telemetry::reload::builder::TracingBuilder;
 use crate::plugins::telemetry::span_factory::SpanMode;
 use crate::plugins::telemetry::tracing::TracingConfigurator;
 use crate::plugins::telemetry::tracing::apollo_telemetry;
