+++ apollo-router/src/plugins/telemetry/tracing/datadog/mod.rs
@@ -7,7 +7,6 @@ use std::fmt::Debug;
 use std::fmt::Formatter;
 use std::time::Duration;
 
-use crate::plugins::telemetry::builder::TracingBuilder;
 use crate::plugins::telemetry::config::{Conf, GenericWith};
 use crate::plugins::telemetry::consts::BUILT_IN_SPAN_NAMES;
 use crate::plugins::telemetry::consts::HTTP_REQUEST_SPAN_NAME;
@@ -20,6 +19,7 @@ use crate::plugins::telemetry::consts::SUBGRAPH_SPAN_NAME;
 use crate::plugins::telemetry::consts::SUPERGRAPH_SPAN_NAME;
 use crate::plugins::telemetry::endpoint::UriEndpoint;
 use crate::plugins::telemetry::otel::named_runtime_channel::NamedTokioRuntime;
+use crate::plugins::telemetry::reload::builder::TracingBuilder;
 use crate::plugins::telemetry::tracing::BatchProcessorConfig;
 use crate::plugins::telemetry::tracing::SpanProcessorExt;
 use crate::plugins::telemetry::tracing::TracingConfigurator;
