+++ apollo-router/src/plugins/telemetry/tracing/mod.rs
@@ -4,8 +4,8 @@ use std::time::Duration;
 
 use super::formatters::APOLLO_CONNECTOR_PREFIX;
 use super::formatters::APOLLO_PRIVATE_PREFIX;
-use crate::plugins::telemetry::builder::TracingBuilder;
 use crate::plugins::telemetry::config::Conf;
+use crate::plugins::telemetry::reload::builder::TracingBuilder;
 use crate::plugins::telemetry::tracing::datadog::DatadogSpanProcessor;
 use opentelemetry::Context;
 use opentelemetry::trace::TraceResult;
