+++ apollo-router/src/plugins/telemetry/tracing/otlp.rs
@@ -1,10 +1,10 @@
 //! Configuration for Otlp tracing.
 use std::result::Result;
 
-use crate::plugins::telemetry::builder::TracingBuilder;
 use crate::plugins::telemetry::config::Conf;
 use crate::plugins::telemetry::otel::named_runtime_channel::NamedTokioRuntime;
 use crate::plugins::telemetry::otlp::TelemetryDataKind;
+use crate::plugins::telemetry::reload::builder::TracingBuilder;
 use crate::plugins::telemetry::tracing::SpanProcessorExt;
 use crate::plugins::telemetry::tracing::TracingConfigurator;
 use opentelemetry_otlp::SpanExporterBuilder;
