+++ apollo-router/src/plugins/telemetry/tracing/zipkin.rs
@@ -1,10 +1,10 @@
 //! Configuration for zipkin tracing.
 use std::sync::LazyLock;
 
-use crate::plugins::telemetry::builder::TracingBuilder;
 use crate::plugins::telemetry::config::{Conf, GenericWith};
 use crate::plugins::telemetry::endpoint::UriEndpoint;
 use crate::plugins::telemetry::otel::named_runtime_channel::NamedTokioRuntime;
+use crate::plugins::telemetry::reload::builder::TracingBuilder;
 use crate::plugins::telemetry::tracing::BatchProcessorConfig;
 use crate::plugins::telemetry::tracing::SpanProcessorExt;
 use crate::plugins::telemetry::tracing::TracingConfigurator;
